+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 - Renamed `*Plume` classes to `*P`, for brevity; for example, use `CollectionsP` instead of `CollectionsPlume`.
 - Removed all deprecated classes and mehods.
 
+## 1.12.0 (2025-??-??)
+
+- Move map-related methods from `CollectionsPlume` to new class `MapsP`
+   * The versions in `CollectionsPlume` are deprecated.
+
 ## 1.11.0 (2025-06-26)
 
 - New class `SIList` for immutable lists that share structure, for efficiency.